macOS - Add flag -march=native when compiling
Add flag -march=native to optimize code specifically for the machine on which the compilation is being performed.
clang compiler then generates code that takes full advantage of the specific features of the processor,
such as special instruction sets (SSE, AVX) which can lead to better performance compared to generic optimizations.